+++ demos/synthetic_signals.py
@@ -6,8 +6,7 @@ from tnmf.utils.demo import SignalTool, st_define_nmf_params
 
 
 @st.cache(hash_funcs={DeltaGenerator: lambda _: None})
-def fit_nmf_model(V, nmf_params, progress_bar):
-    fit_params = {k: nmf_params.pop(k) for k in ('sparsity_H', 'inhibition_strength')}
+def fit_nmf_model(V, nmf_params, fit_params, progress_bar):
     nmf = TransformInvariantNMF(**nmf_params)
     nmf.fit(V, progress_callback=lambda _, x: progress_bar.progress((x + 1) / nmf_params['n_iterations']), **fit_params)
     return nmf
@@ -34,8 +33,8 @@ def main(progress_bar, n_dims: int, verbose: bool = True):
     V, opt_nmf_params = tool.st_generate_input(verbose=verbose)
 
     # define the NMF parameters and fit the model
-    nmf_params = st_define_nmf_params(opt_nmf_params, verbose=verbose)
-    nmf = fit_nmf_model(V, nmf_params, progress_bar)
+    nmf_params, fit_params = st_define_nmf_params(opt_nmf_params, verbose=verbose)
+    nmf = fit_nmf_model(V, nmf_params, fit_params, progress_bar)
 
     # visualize the results
     tool.st_compare_signals(V, nmf.R, verbose=verbose)