+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Unit test Code Coverage Reporting
 - Demo: Racoon image
 - Include examples in pytest bench
+- Support for cross-atom lateral inhibition
 
 ## [0.1.1] - 2021-07-22
 ### Added